Convert data frame of counts to data frame of cases. data frame must contain a column with frequencies (counts) as generated by as.data.frame from a contingency table

Convert data frame of counts to data frame of cases. data frame must contain a column with frequencies (counts) as generated by as.data.frame from a contingency table

Usage

countsToCases(x, countcol = "Freq")

Arguments

x

a data.frame of counts generated from a contingency table.

countcol

character string, name of the column of x containing the counts. Default name of the column is "Freq".

Value

data frame of cases of dimension (total number of counts as sum of "Freq" in x) times 2.

Examples

countsToCases(as.data.frame(HairEyeColor[, , 1]), countcol = "Freq")
#>        Hair   Eye
#> 1     Black Brown
#> 1.1   Black Brown
#> 1.2   Black Brown
#> 1.3   Black Brown
#> 1.4   Black Brown
#> 1.5   Black Brown
#> 1.6   Black Brown
#> 1.7   Black Brown
#> 1.8   Black Brown
#> 1.9   Black Brown
#> 1.10  Black Brown
#> 1.11  Black Brown
#> 1.12  Black Brown
#> 1.13  Black Brown
#> 1.14  Black Brown
#> 1.15  Black Brown
#> 1.16  Black Brown
#> 1.17  Black Brown
#> 1.18  Black Brown
#> 1.19  Black Brown
#> 1.20  Black Brown
#> 1.21  Black Brown
#> 1.22  Black Brown
#> 1.23  Black Brown
#> 1.24  Black Brown
#> 1.25  Black Brown
#> 1.26  Black Brown
#> 1.27  Black Brown
#> 1.28  Black Brown
#> 1.29  Black Brown
#> 1.30  Black Brown
#> 1.31  Black Brown
#> 2     Brown Brown
#> 2.1   Brown Brown
#> 2.2   Brown Brown
#> 2.3   Brown Brown
#> 2.4   Brown Brown
#> 2.5   Brown Brown
#> 2.6   Brown Brown
#> 2.7   Brown Brown
#> 2.8   Brown Brown
#> 2.9   Brown Brown
#> 2.10  Brown Brown
#> 2.11  Brown Brown
#> 2.12  Brown Brown
#> 2.13  Brown Brown
#> 2.14  Brown Brown
#> 2.15  Brown Brown
#> 2.16  Brown Brown
#> 2.17  Brown Brown
#> 2.18  Brown Brown
#> 2.19  Brown Brown
#> 2.20  Brown Brown
#> 2.21  Brown Brown
#> 2.22  Brown Brown
#> 2.23  Brown Brown
#> 2.24  Brown Brown
#> 2.25  Brown Brown
#> 2.26  Brown Brown
#> 2.27  Brown Brown
#> 2.28  Brown Brown
#> 2.29  Brown Brown
#> 2.30  Brown Brown
#> 2.31  Brown Brown
#> 2.32  Brown Brown
#> 2.33  Brown Brown
#> 2.34  Brown Brown
#> 2.35  Brown Brown
#> 2.36  Brown Brown
#> 2.37  Brown Brown
#> 2.38  Brown Brown
#> 2.39  Brown Brown
#> 2.40  Brown Brown
#> 2.41  Brown Brown
#> 2.42  Brown Brown
#> 2.43  Brown Brown
#> 2.44  Brown Brown
#> 2.45  Brown Brown
#> 2.46  Brown Brown
#> 2.47  Brown Brown
#> 2.48  Brown Brown
#> 2.49  Brown Brown
#> 2.50  Brown Brown
#> 2.51  Brown Brown
#> 2.52  Brown Brown
#> 3       Red Brown
#> 3.1     Red Brown
#> 3.2     Red Brown
#> 3.3     Red Brown
#> 3.4     Red Brown
#> 3.5     Red Brown
#> 3.6     Red Brown
#> 3.7     Red Brown
#> 3.8     Red Brown
#> 3.9     Red Brown
#> 4     Blond Brown
#> 4.1   Blond Brown
#> 4.2   Blond Brown
#> 5     Black  Blue
#> 5.1   Black  Blue
#> 5.2   Black  Blue
#> 5.3   Black  Blue
#> 5.4   Black  Blue
#> 5.5   Black  Blue
#> 5.6   Black  Blue
#> 5.7   Black  Blue
#> 5.8   Black  Blue
#> 5.9   Black  Blue
#> 5.10  Black  Blue
#> 6     Brown  Blue
#> 6.1   Brown  Blue
#> 6.2   Brown  Blue
#> 6.3   Brown  Blue
#> 6.4   Brown  Blue
#> 6.5   Brown  Blue
#> 6.6   Brown  Blue
#> 6.7   Brown  Blue
#> 6.8   Brown  Blue
#> 6.9   Brown  Blue
#> 6.10  Brown  Blue
#> 6.11  Brown  Blue
#> 6.12  Brown  Blue
#> 6.13  Brown  Blue
#> 6.14  Brown  Blue
#> 6.15  Brown  Blue
#> 6.16  Brown  Blue
#> 6.17  Brown  Blue
#> 6.18  Brown  Blue
#> 6.19  Brown  Blue
#> 6.20  Brown  Blue
#> 6.21  Brown  Blue
#> 6.22  Brown  Blue
#> 6.23  Brown  Blue
#> 6.24  Brown  Blue
#> 6.25  Brown  Blue
#> 6.26  Brown  Blue
#> 6.27  Brown  Blue
#> 6.28  Brown  Blue
#> 6.29  Brown  Blue
#> 6.30  Brown  Blue
#> 6.31  Brown  Blue
#> 6.32  Brown  Blue
#> 6.33  Brown  Blue
#> 6.34  Brown  Blue
#> 6.35  Brown  Blue
#> 6.36  Brown  Blue
#> 6.37  Brown  Blue
#> 6.38  Brown  Blue
#> 6.39  Brown  Blue
#> 6.40  Brown  Blue
#> 6.41  Brown  Blue
#> 6.42  Brown  Blue
#> 6.43  Brown  Blue
#> 6.44  Brown  Blue
#> 6.45  Brown  Blue
#> 6.46  Brown  Blue
#> 6.47  Brown  Blue
#> 6.48  Brown  Blue
#> 6.49  Brown  Blue
#> 7       Red  Blue
#> 7.1     Red  Blue
#> 7.2     Red  Blue
#> 7.3     Red  Blue
#> 7.4     Red  Blue
#> 7.5     Red  Blue
#> 7.6     Red  Blue
#> 7.7     Red  Blue
#> 7.8     Red  Blue
#> 7.9     Red  Blue
#> 8     Blond  Blue
#> 8.1   Blond  Blue
#> 8.2   Blond  Blue
#> 8.3   Blond  Blue
#> 8.4   Blond  Blue
#> 8.5   Blond  Blue
#> 8.6   Blond  Blue
#> 8.7   Blond  Blue
#> 8.8   Blond  Blue
#> 8.9   Blond  Blue
#> 8.10  Blond  Blue
#> 8.11  Blond  Blue
#> 8.12  Blond  Blue
#> 8.13  Blond  Blue
#> 8.14  Blond  Blue
#> 8.15  Blond  Blue
#> 8.16  Blond  Blue
#> 8.17  Blond  Blue
#> 8.18  Blond  Blue
#> 8.19  Blond  Blue
#> 8.20  Blond  Blue
#> 8.21  Blond  Blue
#> 8.22  Blond  Blue
#> 8.23  Blond  Blue
#> 8.24  Blond  Blue
#> 8.25  Blond  Blue
#> 8.26  Blond  Blue
#> 8.27  Blond  Blue
#> 8.28  Blond  Blue
#> 8.29  Blond  Blue
#> 9     Black Hazel
#> 9.1   Black Hazel
#> 9.2   Black Hazel
#> 9.3   Black Hazel
#> 9.4   Black Hazel
#> 9.5   Black Hazel
#> 9.6   Black Hazel
#> 9.7   Black Hazel
#> 9.8   Black Hazel
#> 9.9   Black Hazel
#> 10    Brown Hazel
#> 10.1  Brown Hazel
#> 10.2  Brown Hazel
#> 10.3  Brown Hazel
#> 10.4  Brown Hazel
#> 10.5  Brown Hazel
#> 10.6  Brown Hazel
#> 10.7  Brown Hazel
#> 10.8  Brown Hazel
#> 10.9  Brown Hazel
#> 10.10 Brown Hazel
#> 10.11 Brown Hazel
#> 10.12 Brown Hazel
#> 10.13 Brown Hazel
#> 10.14 Brown Hazel
#> 10.15 Brown Hazel
#> 10.16 Brown Hazel
#> 10.17 Brown Hazel
#> 10.18 Brown Hazel
#> 10.19 Brown Hazel
#> 10.20 Brown Hazel
#> 10.21 Brown Hazel
#> 10.22 Brown Hazel
#> 10.23 Brown Hazel
#> 10.24 Brown Hazel
#> 11      Red Hazel
#> 11.1    Red Hazel
#> 11.2    Red Hazel
#> 11.3    Red Hazel
#> 11.4    Red Hazel
#> 11.5    Red Hazel
#> 11.6    Red Hazel
#> 12    Blond Hazel
#> 12.1  Blond Hazel
#> 12.2  Blond Hazel
#> 12.3  Blond Hazel
#> 12.4  Blond Hazel
#> 13    Black Green
#> 13.1  Black Green
#> 13.2  Black Green
#> 14    Brown Green
#> 14.1  Brown Green
#> 14.2  Brown Green
#> 14.3  Brown Green
#> 14.4  Brown Green
#> 14.5  Brown Green
#> 14.6  Brown Green
#> 14.7  Brown Green
#> 14.8  Brown Green
#> 14.9  Brown Green
#> 14.10 Brown Green
#> 14.11 Brown Green
#> 14.12 Brown Green
#> 14.13 Brown Green
#> 14.14 Brown Green
#> 15      Red Green
#> 15.1    Red Green
#> 15.2    Red Green
#> 15.3    Red Green
#> 15.4    Red Green
#> 15.5    Red Green
#> 15.6    Red Green
#> 16    Blond Green
#> 16.1  Blond Green
#> 16.2  Blond Green
#> 16.3  Blond Green
#> 16.4  Blond Green
#> 16.5  Blond Green
#> 16.6  Blond Green
#> 16.7  Blond Green