Selects columns defined by characters varsample and varfactor from a data.frame

Selects columns defined by characters varsample and varfactor from dataframe, returns selected columns with their names.

get_samples_fact_inputfile(dataframe, varsample, varfactor)

Arguments

dataframe
data.frame or list containing at least two columns with column headings of data type character. Data must be column wise ordered.

varsample
column name of dependent variable in dataframe, datatype character

varfactor
column name of independent variable in dataframe, datatype character

Value

selected columns, sample, factor, name_of_sample (character string equaling varsample), name_of_factor (character string equaling varsample)

Examples

get_samples_fact_inputfile(trees,"Girth","Height")

#> $samples
#>  [1]  8.3  8.6  8.8 10.5 10.7 10.8 11.0 11.0 11.1 11.2 11.3 11.4 11.4 11.7 12.0
#> [16] 12.9 12.9 13.3 13.7 13.8 14.0 14.2 14.5 16.0 16.3 17.3 17.5 17.9 18.0 18.0
#> [31] 20.6
#> 
#> $fact
#>  [1] 70 65 63 72 81 83 66 75 80 75 79 76 76 69 75 74 85 86 71 64 78 80 74 72 77
#> [26] 81 82 80 80 80 87
#> 
#> $name_of_sample
#> [1] "Girth"
#> 
#> $name_of_factor
#> [1] "Height"
#> 
#> $matchingCriteria
#> [1] ""
#>